\name{edbGetActiveParticipants}
\alias{edbGetActiveParticipants}
\title{Returns dataframe with info about all active participants}
\description{
Returns dataframe with info about all active participants
}
\usage{edbGetActiveParticipants(d)}
\arguments{
  \item{d}{cleaned enrollment database as dataframe}
}
\value{Returns a dataframe}
\author{John J. Curtin \email{jjcurtin@wisc.edu}}