\name{edbGetCleanDB}
\alias{edbGetCleanDB}
\title{Download a clean Qualtrics enrollment database}
\description{
Downloads, cleans, and verifies a Qualtrics enrollment database
}
\usage{edbGetCleanDB(SurveyID)}
\arguments{
  \item{SurveyID}{Qualtrics SurveyID}
}
\value{returns a cleaned dataframe including all enrollment events}
\author{John J. Curtin \email{jjcurtin@wisc.edu}}