\name{edbGetNextSubID}
\alias{edbGetNextSubID}
\title{Get Next Unassigned SubID}
\description{
Returns the next consecutive unassigned SubID
}
\usage{edbGetNextSubID(SurveyID)}
\arguments{
  \item{SurveyID}{Qualtrics SurveyID}
}
\value{Returns the next unassigned SubID in the console}
\author{John J. Curtin \email{jjcurtin@wisc.edu}}