\name{edbGetRawDB}
\alias{edbGetRawDB}
\title{Gets raw enrollment databasefrom Qualtrics}
\description{
Downloads a dataframe with all the raw responses in the Qualtrics enrollment database
}
\usage{edbGetRawDB(SurveyID)}
\arguments{
  \item{SurveyID}{Qualtrics SurveyID}
}
\value{Returns dataframe with raw records from qualtrics enrollment database}
\author{John J. Curtin \email{jjcurtin@wisc.edu}}