\name{edbQualtricsInput}
\alias{edbQualtricsInput}
\title{Opens Qualtrics Survey for input in browswer}
\description{
Opens Qualtrics survey for input in browser along with StaffID tag
}
\usage{edbQualtricsInput(StaffID, URL)}
\arguments{
  \item{StaffID}{String with StaffID.  LastNameFirstInitial is our default}
  \item{URL}{Sting with URL to the survey}
}
\value{No return value}
\author{John J. Curtin \email{jjcurtin@wisc.edu}}