\name{edbSummaryStatus}
\alias{edbSummaryStatus}
\title{Summary of Disposition Status (total and by Sex)}
\description{
Returns a dataframe with number of participants by each disposition (and Sex)
}
\usage{edbSummaryStatus(d)}
\arguments{
  \item{d}{A cleaned enrollment database}
}
\value{Returns a dataframe with number of participants by each disposition (and Sex)}
\author{John J. Curtin \email{jjcurtin@wisc.edu}}