\name{rdbFormat}
\alias{rdbFormat}
\title{Reduces variables to faciliate easier display with View}
\description{
Reduces variables to faciliate easier display with View
} 
\usage{rdbFormat(d, Verbose=FALSE)}
\arguments{
  \item{d}{A dataframe with recruiting database or call list}
  \item{Verbose}{View more (TRUE) or fewer (FALSE) columns of the database; Default=FALSE}
}
\value{dataframe with reduced number of columns}
\author{John J. Curtin \email{jjcurtin@wisc.edu}}