\name{rdbGetRawDB}
\alias{rdbGetRawDB}
\title{Gets raw database from Qualtrics}
\description{
Downloads a dataframe with all the raw responses in the Qualtrics database
}
\usage{rdbGetRawDB(SurveyID)}
\arguments{
  \item{SurveyID}{Qualtrics SurveyID}
}
\value{Returns dataframe with raw records from qualtrics database}
\author{John J. Curtin \email{jjcurtin@wisc.edu}}