\name{rdbGetSummary}
\alias{rdbGetSummary}
\title{Provides summary statistics and graphs about recruiting progress}
\description{
Provides a wealth of info to help track progress on recruiting in a pdf
}
\usage{rdbGetSummary(d, FullFileName)}
\arguments{
  \item{d}{cleaned recruiting database as dataframe}
  \item{FullFileName}{String with path and filename for pdf output}
}
\value{No return value.  Creates a pdf saved to disk}
\author{John J. Curtin \email{jjcurtin@wisc.edu}}